Separate Business logic from data access layer; some minor fixes
