+++ src/database/populate_db.py
@@ -4,26 +4,25 @@ import asyncio
 from uuid import uuid4
 
 import pandas as pd
-from tqdm import tqdm
 
 from src.config import settings
 from src.routers.schemas import DocumentSchema
-from src.services.repository import Repository
+from src.services.documents import DocumentsService
 
-# Определяем путь к вашему CSV файлу
 CSV_PATH = "data/raw/evidently_table.csv"
 
 
 async def populate_database() -> None:
     df = pd.read_csv(CSV_PATH)
 
-    repository = await Repository.create(
+    documents_service = DocumentsService(
         collection_name=settings.COLLECTION_NAME, model_name=settings.EMBEDDING_MODEL_NAME
     )
 
     print("Populating database...")
 
-    for index, row in tqdm(df.iterrows()):
+    docs_data = []
+    for index, row in df.iterrows():
         doc_data = DocumentSchema(
             doc_id=uuid4(),
             company=row.get("Company"),
@@ -39,8 +38,9 @@ async def populate_database() -> None:
             score=0.0,
             metadata=None,
         )
+        docs_data.append(doc_data)
 
-        await repository.add_document(doc_data)
+    await documents_service.add_documents(docs_data=docs_data)
 
 
 if __name__ == "__main__":
