+++ src/repositories/__init__.py
@@ -1,0 +1,2 @@
+from .collection import CollectionRepository  # noqa: F401
+from .docs import DocsRepository  # noqa: F401
